+++ objects/baza/lambda.rb
@@ -34,6 +34,7 @@ require 'net/scp'
 require 'openssl'
 require_relative 'ec2'
 require_relative 'shell'
+require_relative 'image'
 
 # Function in AWS Lambda.
 #
@@ -53,6 +54,7 @@ class Baza::Lambda
   def initialize(humans, account, key, secret, region, sgroup, subnet, image, ssh,
     tbot: Baza::Tbot::Fake.new, loog: Loog::NULL, type: 't2.xlarge', user: 'ubuntu', port: 22)
     @humans = humans
+    @image = Baza::Image.new(humans, account, region, loog:)
     @ec2 = Baza::EC2.new(key, secret, region, sgroup, subnet, image, type:, loog:)
     @shell = Baza::Shell.new(ssh, user, port, loog:) unless key.empty?
     @account = account
@@ -71,7 +73,7 @@ class Baza::Lambda
   def deploy(tag = 'latest')
     return unless dirty?
     Dir.mktmpdir do |home|
-      zip = pack(File.join(home, 'image.zip'))
+      zip = @image.pack(File.join(home, 'image.zip'))
       sha = Digest::SHA1.hexdigest(File.binread(zip))
       break if aws_sha == sha
       instance_id = @ec2.run_instance
@@ -140,90 +142,6 @@ class Baza::Lambda
   # @return [String] The SHA or '' if no lambda function found
   def aws_sha
     ''
-  end
-
-  # Package all necessary files for Docker image.
-  #
-  # @param [String] file Path of the .zip file to create
-  # @return [String] File path of .zip
-  def pack(file)
-    Dir.mktmpdir do |home|
-      [
-        '../../Gemfile',
-        '../../Gemfile.lock',
-        '../../assets/lambda/entry.rb',
-        '../../assets/lambda/install-pgsql.sh'
-      ].each do |f|
-        FileUtils.copy(File.join(__dir__, f), File.join(home, File.basename(f)))
-      end
-      FileUtils.mkdir_p(File.join(home, 'swarms'))
-      File.write(File.join(home, 'swarms/.keep'), '')
-      installs = []
-      each_swarm do |swarm|
-        dir = checkout(swarm)
-        next if dir.nil?
-        sub = "swarms/#{swarm.name}"
-        target = File.join(home, sub)
-        FileUtils.mkdir_p(File.dirname(target))
-        FileUtils.copy_entry(dir, target)
-        installs << install(target, sub)
-      end
-      dockerfile = Liquid::Template.parse(File.read(File.join(__dir__, '../../assets/lambda/Dockerfile'))).render(
-        'from' => "#{@account}.dkr.ecr.#{@region}.amazonaws.com/zerocracy/baza:basic",
-        'installs' => installs.join("\n")
-      )
-      File.write(File.join(home, 'Dockerfile'), dockerfile)
-      @loog.debug("This is the Dockerfile:\n#{dockerfile}")
-      Baza::Zip.new(file, loog: @loog).pack(home)
-    end
-    file
-  end
-
-  # Create install commands for Docker, from this directory.
-  #
-  # @param [String] dir The local directory with swarm content files, e.g. "/tmp/bar/foo-contents"
-  # @param [String] sub Subdirectory inside docker image, e.g. "swarms/foo"
-  def install(dir, sub)
-    gemfile = File.join(dir, 'Gemfile.lock')
-    if File.exist?(gemfile)
-      "RUN bundle install --gemfile=#{sub}/Gemfile"
-    else
-      ''
-    end
-  end
-
-  # Checkout swarm and return the directory where it's located. Also,
-  # update its SHA if necessary.
-  #
-  # @param [Baza::Swarm] swarm The swarm
-  # @return [String] Path to location
-  def checkout(swarm)
-    elapsed(@loog, intro: "Checked out #{swarm.name} swarm") do
-      sub = "swarms/#{swarm.name}"
-      dir = File.join('/tmp', sub)
-      FileUtils.mkdir_p(File.dirname(dir))
-      git = ['set -ex', 'date', 'git --version']
-      if File.exist?(dir)
-        git += ["cd #{dir}", 'git pull']
-      else
-        git << "git clone -b #{swarm.branch} --depth=1 --single-branch git@github.com:#{swarm.repository}.git #{dir}"
-      end
-      git << 'git rev-parse HEAD'
-      stdout = `(#{git.join(' && ')}) 2>&1`
-      @loog.debug("Checkout log of #{swarm.name}:\n#{stdout}")
-      swarm.stdout!(stdout)
-      code = $CHILD_STATUS.exitstatus
-      swarm.exit!(code)
-      return nil unless code.zero?
-      dir
-    end
-  end
-
-  # Iterate all swarms that need to be deployed.
-  def each_swarm
-    @humans.pgsql.exec('SELECT * FROM swarm').each do |row|
-      yield @humans.find_swarm(row['repository'], row['branch'])
-    end
   end
 
   # Returns TRUE if at least one swarm is "dirty" and because of that
